requirements.txt
torch
transformers
peft
datasets
PyYAML
accelerate
bitsandbytes